RHCLOUD-27242 Move SEDA logic to connector-common module
