+++ connector-common/src/main/java/com/redhat/cloud/notifications/connector/ConnectorConfig.java
@@ -26,6 +26,9 @@ public class ConnectorConfig {
     private static final String REDELIVERY_COUNTER_NAME = "notifications.connector.redelivery.counter-name";
     private static final String REDELIVERY_DELAY = "notifications.connector.redelivery.delay";
     private static final String REDELIVERY_MAX_ATTEMPTS = "notifications.connector.redelivery.max-attempts";
+    private static final String SEDA_CONCURRENT_CONSUMERS = "notifications.connector.seda.concurrent-consumers";
+    private static final String SEDA_ENABLED = "notifications.connector.seda.enabled";
+    private static final String SEDA_QUEUE_SIZE = "notifications.connector.seda.queue-size";
 
     @ConfigProperty(name = ENDPOINT_CACHE_MAX_SIZE, defaultValue = "100")
     int endpointCacheMaxSize;
@@ -62,6 +65,17 @@ public class ConnectorConfig {
     @ConfigProperty(name = REDELIVERY_MAX_ATTEMPTS, defaultValue = "2")
     int redeliveryMaxAttempts;
 
+    // https://camel.apache.org/components/4.0.x/seda-component.html#_component_option_concurrentConsumers
+    @ConfigProperty(name = SEDA_CONCURRENT_CONSUMERS, defaultValue = "1")
+    int sedaConcurrentConsumers;
+
+    @ConfigProperty(name = SEDA_ENABLED, defaultValue = "false")
+    boolean sedaEnabled;
+
+    // https://camel.apache.org/components/4.0.x/seda-component.html#_component_option_queueSize
+    @ConfigProperty(name = SEDA_QUEUE_SIZE, defaultValue = "1000")
+    int sedaQueueSize;
+
     public void log() {
         log(Collections.emptyMap());
     }
@@ -80,6 +94,9 @@ public class ConnectorConfig {
         config.put(REDELIVERY_COUNTER_NAME, redeliveryCounterName);
         config.put(REDELIVERY_DELAY, redeliveryDelay);
         config.put(REDELIVERY_MAX_ATTEMPTS, redeliveryMaxAttempts);
+        config.put(SEDA_CONCURRENT_CONSUMERS, sedaConcurrentConsumers);
+        config.put(SEDA_ENABLED, sedaEnabled);
+        config.put(SEDA_QUEUE_SIZE, sedaQueueSize);
         config.putAll(additionalConfig);
 
         Log.info("=== Connector configuration ===");
@@ -130,5 +147,17 @@ public class ConnectorConfig {
 
     public int getRedeliveryMaxAttempts() {
         return redeliveryMaxAttempts;
+    }
+
+    public int getSedaConcurrentConsumers() {
+        return sedaConcurrentConsumers;
+    }
+
+    public boolean isSedaEnabled() {
+        return sedaEnabled;
+    }
+
+    public int getSedaQueueSize() {
+        return sedaQueueSize;
     }
 }
