+++ connector-common/src/main/java/com/redhat/cloud/notifications/connector/EngineToConnectorRouteBuilder.java
@@ -2,6 +2,7 @@ package com.redhat.cloud.notifications.connector;
 
 import jakarta.inject.Inject;
 import org.apache.camel.builder.endpoint.EndpointRouteBuilder;
+import org.apache.camel.component.seda.SedaComponent;
 
 import static org.apache.camel.LoggingLevel.DEBUG;
 import static org.apache.camel.builder.endpoint.dsl.KafkaEndpointBuilderFactory.KafkaEndpointConsumerBuilder;
@@ -45,6 +46,10 @@ public abstract class EngineToConnectorRouteBuilder extends EndpointRouteBuilder
                 .handled(true)
                 .process(exceptionProcessor);
 
+        if (connectorConfig.isSedaEnabled()) {
+            configureSedaComponent();
+        }
+
         from(buildKafkaEndpoint())
                 .routeId(ENGINE_TO_CONNECTOR)
                 .to(log(getClass().getName()).level("DEBUG").showHeaders(true).showBody(true))
@@ -53,7 +58,14 @@ public abstract class EngineToConnectorRouteBuilder extends EndpointRouteBuilder
                 .removeHeaders("*")
                 .process(incomingCloudEventProcessor)
                 .to(log(getClass().getName()).level("DEBUG").showProperties(true))
-                .to(direct(ENGINE_TO_CONNECTOR));
+                // TODO The following lines should be removed when all connectors are migrated to SEDA.
+                .choice()
+                .when(exchange -> connectorConfig.isSedaEnabled())
+                .to(seda(ENGINE_TO_CONNECTOR))
+                .endChoice()
+                .otherwise()
+                .to(direct(ENGINE_TO_CONNECTOR))
+                .end();
 
         configureRoute();
     }
@@ -66,5 +78,13 @@ public abstract class EngineToConnectorRouteBuilder extends EndpointRouteBuilder
                 .maxPollRecords(connectorConfig.getIncomingKafkaMaxPollRecords())
                 .maxPollIntervalMs(connectorConfig.getIncomingKafkaMaxPollIntervalMs())
                 .pollOnError(connectorConfig.getIncomingKafkaPollOnError());
+    }
+
+    private void configureSedaComponent() {
+        SedaComponent component = getContext().getComponent("seda", SedaComponent.class);
+        component.setConcurrentConsumers(connectorConfig.getSedaConcurrentConsumers());
+        component.setQueueSize(connectorConfig.getSedaQueueSize());
+        // The Kafka messages consumption is blocked (paused) when the SEDA queue is full.
+        component.setDefaultBlockWhenFull(true);
     }
 }
