+++ connector-webhook/src/main/java/com/redhat/cloud/notifications/connector/webhook/WebhookRouteBuilder.java
@@ -8,7 +8,6 @@ import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import org.apache.camel.builder.endpoint.dsl.HttpEndpointBuilderFactory;
 import org.apache.camel.component.http.HttpComponent;
-import org.apache.camel.component.seda.SedaComponent;
 import org.apache.hc.core5.util.Timeout;
 import org.apache.http.conn.ssl.NoopHostnameVerifier;
 
@@ -26,13 +25,11 @@ import static com.redhat.cloud.notifications.connector.webhook.SslTrustAllManage
 import static org.apache.camel.Exchange.CONTENT_TYPE;
 import static org.apache.camel.LoggingLevel.INFO;
 
-
 @ApplicationScoped
 public class WebhookRouteBuilder extends EngineToConnectorRouteBuilder {
 
     public static final String CLOUD_EVENT_TYPE_PREFIX = "com.redhat.console.notification.toCamel.";
     private static final String APPLICATION_JSON = "application/json";
-    private static final String SEDA = "seda";
 
     @Inject
     WebhookConnectorConfig webhookConnectorConfig;
@@ -42,12 +39,8 @@ public class WebhookRouteBuilder extends EngineToConnectorRouteBuilder {
 
         configureHttpComponent("http");
         configureHttpComponent("https");
-        configureSedaComponent();
 
-        from(direct(ENGINE_TO_CONNECTOR))
-            .to(seda(SEDA));
-
-        from(seda(SEDA))
+        from(seda(ENGINE_TO_CONNECTOR))
             .setHeader(CONTENT_TYPE, constant(APPLICATION_JSON))
             .routeId(webhookConnectorConfig.getConnectorName())
             .choice()
@@ -81,13 +74,6 @@ public class WebhookRouteBuilder extends EngineToConnectorRouteBuilder {
         component.setConnectionsPerRoute(webhookConnectorConfig.getHttpConnectionsPerRoute());
         component.setMaxTotalConnections(webhookConnectorConfig.getHttpMaxTotalConnections());
         component.setFollowRedirects(true);
-    }
-
-    private void configureSedaComponent() {
-        SedaComponent component = getContext().getComponent("seda", SedaComponent.class);
-        component.setConcurrentConsumers(webhookConnectorConfig.getSedaConcurrentConsumers());
-        component.setQueueSize(webhookConnectorConfig.getSedaQueueSize());
-        component.setDefaultBlockWhenFull(true);
     }
 
     private HttpEndpointBuilderFactory.HttpEndpointBuilder buildUnsecureSslEndpoint() {
