+++ connector-webhook/src/main/resources/application.properties
@@ -4,6 +4,11 @@ notifications.connector.kafka.outgoing.topic=${mp.messaging.fromcamel.topic}
 notifications.connector.name=webhook
 notifications.connector.redelivery.counter-name=camel.webhook.retry.counter
 notifications.connector.alternative.names=ansible
+# The following value matches the default value of the `connectionsPerRoute` option from the Camel `http` component.
+notifications.connector.seda.concurrent-consumers=20
+notifications.connector.seda.enabled=true
+# The following value matches the default value of the `connectionsPerRoute` option from the Camel `http` component.
+notifications.connector.seda.queue-size=20
 
 quarkus.http.port=9006
